@@ -25,53 +25,57 @@
     "fmt": "prettier --write \"src/**/*.{js,json,css}\""
   },
   "dependencies": {
-    "@babel/core": "7.3.3",
-    "@babel/runtime": "7.3.1",
-    "@dlghq/babel-preset-dialog": "4.3.4",
-    "@dlghq/postcss-dialog": "1.13.2",
-    "babel-loader": "8.0.5",
-    "bluebird": "3.5.3",
+    "@babel/core": "7.4.5",
+    "@dlghq/babel-preset-dialog": "4.4.0",
+    "@dlghq/postcss-dialog": "1.15.0",
+    "astroturf": "0.9.6",
+    "babel-loader": "8.0.6",
+    "bluebird": "3.5.5",
     "build-number": "1.0.0",
     "chalk": "2.4.2",
-    "commander": "2.19.0",
+    "commander": "2.20.0",
     "compression-webpack-plugin": "2.0.0",
-    "copy-webpack-plugin": "4.6.0",
-    "css-loader": "1.0.1",
+    "copy-webpack-plugin": "5.0.3",
+    "css-loader": "3.0.0",
+    "cssnano": "4.1.10",
     "duplicate-package-checker-webpack-plugin": "3.0.0",
     "electron-builder": "20.38.5",
     "electron-publisher-s3": "20.17.2",
-    "file-loader": "2.0.0",
+    "file-loader": "3.0.1",
     "generate-json-webpack-plugin": "0.3.1",
     "html-webpack-plugin": "3.2.0",
     "json-loader": "0.5.7",
     "lodash": "4.17.11",
-    "mini-css-extract-plugin": "0.4.5",
-    "optimize-css-assets-webpack-plugin": "5.0.1",
+    "mini-css-extract-plugin": "0.7.0",
+    "optimize-css-assets-webpack-plugin": "5.0.3",
+    "postcss-css-variables": "0.13.0",
+    "postcss-import": "12.0.1",
     "postcss-loader": "3.0.0",
-    "progress-bar-webpack-plugin": "1.11.0",
+    "postcss-preset-env": "6.6.0",
+    "progress-bar-webpack-plugin": "1.12.1",
     "raw-loader": "0.5.1",
     "style-loader": "0.23.1",
-    "svg-sprite-loader": "4.1.3",
-    "terser-webpack-plugin": "1.1.0",
+    "svg-sprite-loader": "4.1.6",
+    "terser-webpack-plugin": "1.3.0",
     "url-loader": "1.1.2",
     "val-loader": "1.1.1",
-    "webpack": "4.26.1",
-    "webpack-dev-server": "3.1.10",
+    "webpack": "4.35.0",
+    "webpack-dev-server": "3.7.2",
     "webpack-sentry-plugin": "1.16.0",
     "worker-loader": "2.0.0",
     "yml-loader": "2.1.0"
   },
   "devDependencies": {
-    "@babel/cli": "7.2.3",
+    "@babel/cli": "7.4.4",
     "@babel/plugin-syntax-flow": "7.2.0",
-    "@babel/plugin-transform-flow-comments": "7.2.0",
-    "@babel/plugin-transform-flow-strip-types": "7.2.3",
-    "@babel/plugin-transform-runtime": "7.2.0",
-    "@babel/preset-env": "7.3.1",
-    "@dlghq/eslint-config-dialog": "3.4.0",
-    "eslint": "5.9.0",
-    "eslint-config-prettier": "3.3.0",
-    "flow-bin": "0.87.0",
-    "prettier": "1.15.2"
+    "@babel/plugin-transform-flow-comments": "7.4.4",
+    "@babel/plugin-transform-flow-strip-types": "7.4.4",
+    "@babel/plugin-transform-runtime": "7.4.4",
+    "@babel/preset-env": "7.4.5",
+    "@dlghq/eslint-config-dialog": "4.0.8",
+    "eslint": "5.16.0",
+    "eslint-config-prettier": "3.6.0",
+    "flow-bin": "0.100.0",
+    "prettier": "1.18.2"
   }
 }
